fixed errors

+++ routes/api/contacts.js
@@ -1,19 +1,18 @@
 const express = require("express");
 
 const CreateError = require("http-errors");
+const { Mongoose } = require("mongoose");
 const router = express.Router();
 
 const { Contact, schemas } = require("../../models/contact");
 
 router.get("/", async (req, res, next) => {
   try {
-    console.log(`start`);
     const result = await Contact.find({}, "-createdAt -updatedAt");
     if (!result) {
       throw new CreateError(404, "Not found");
     }
     res.json(result);
-    console.log(`rezult`);
   } catch (error) {
     next(error);
   }
@@ -22,6 +21,10 @@ router.get("/", async (req, res, next) => {
 router.get("/:contactId", async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!Mongoose.Types.ObjectId.isValid(contactId)) {
+      throw new CreateError(404, "Not valid ID");
+    }
+
     const result = await Contact.findById(contactId);
     if (!result) {
       throw new CreateError(404, "Not found");
@@ -29,9 +32,6 @@ router.get("/:contactId", async (req, res, next) => {
 
     res.json(result);
   } catch (error) {
-    if (error.message.includes("Cast to ObjectId")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
@@ -50,9 +50,6 @@ router.post("/", async (req, res, next) => {
 
     res.json(result);
   } catch (error) {
-    if (error.message.includes("validation failed")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
@@ -60,15 +57,17 @@ router.post("/", async (req, res, next) => {
 router.delete("/:contactId", async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!Mongoose.Types.ObjectId.isValid(contactId)) {
+      throw new CreateError(404, "Not valid ID");
+    }
+
     const result = await Contact.findByIdAndDelete(contactId);
     if (!result) {
       throw new CreateError(404, "Not found");
     }
+
     res.json({ message: "Contact deleted" });
   } catch (error) {
-    if (error.message.includes("Cast to ObjectId")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
@@ -76,6 +75,9 @@ router.delete("/:contactId", async (req, res, next) => {
 router.put("/:contactId", async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!Mongoose.Types.ObjectId.isValid(contactId)) {
+      throw new CreateError(404, "Not valid ID");
+    }
     const { error } = await schemas.add.validate(req.body);
     if (error) {
       throw new CreateError(400, error.message);
@@ -90,9 +92,6 @@ router.put("/:contactId", async (req, res, next) => {
 
     res.json(result);
   } catch (error) {
-    if (error.message.includes("Cast to ObjectId")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
@@ -100,6 +99,10 @@ router.put("/:contactId", async (req, res, next) => {
 router.patch("/:contactId/favorite", async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!Mongoose.Types.ObjectId.isValid(contactId)) {
+      throw new CreateError(404, "Not valid ID");
+    }
+
     const { error } = await schemas.updateFavorite.validate(req.body);
     if (error) {
       throw new CreateError(400, error.message);
@@ -114,9 +117,6 @@ router.patch("/:contactId/favorite", async (req, res, next) => {
 
     res.json(result);
   } catch (error) {
-    if (error.message.includes("Cast to ObjectId")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
